Renamed SalaryInput component
