+++ src/components/ui/Form/Input/AmountFormatInput.vue
